Debugging and fixing bug with random game over.
I was debugging this issue that even though I have removed customer from the scene using `remove` function, it was still moving and was causing a game over when it got too close to the player.

I have tried to fix that problem by explicitly removing interval call using `clearInterval` and it helped.
However when the object moves meanwhile (because timings for intervals are different), it doesn't search for object correctly and interval is still not removed and same bug appears (just not so often). So I'll have to fix that later.