Finally fix the bug. Add `addSpriteWithReturn` function with a bit hacky solution.
I was still debugging the same bug, tried different solutions and finally found a solution (I hope).
Instead of making a custom interval for each customer I made an interval that moves all customers at the "same time". It uses `getAll` function which now correctly checks whether a customer wasn't removed from the scene (which was the bug before).
This also simplified the code a bit.